moved the Hypermc project into the src folder
