--- Hypermc/Mod.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Hypermc;
-
-namespace Hypermc
-{
-	public class Mod
-	{
-		public string Name { get; }
-
-		public Mod(string name)
-		{
-			Name = name;
-		}
-	}
-}
