+++ src/HyperMC/Data/ModpackData.cs
