+++ src/HyperMC/HyperMcView.resx
