+++ src/HyperMC/Mod.cs
@@ -1,7 +1,19 @@
-﻿namespace Hypermc
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Hypermc;
+
+namespace Hypermc
 {
-    public class Mod
-    {
-        public string Name { get; init; }
-    }
+	public class Mod
+	{
+		public string Name { get; }
+
+		public Mod(string name)
+		{
+			Name = name;
+		}
+	}
 }
