+++ src/HyperMC/Properties/Resources.resx
