+++ src/HyperMC/Resources/DefaultModpackImage.png
