+++ src/HyperMC/Resources/logo.png
