+++ src/HyperMC/UI/Dialogs/CreateModpackDialog.cs
