+++ src/HyperMC/UI/UserControls/ModpackBox.resx
