+++ src/HyperMC/UI/Views/IView.cs
