+++ src/HyperMC/UI/Views/IViewHost.cs
