+++ src/HyperMC/UI/Views/SettingView.cs
