+++ src/HyperMC/UI/Views/SettingView.resx
