+++ src/HyperMC/Utility/MessageDialog.cs
