+++ src/HyperMC/Utility/Utils.cs
