+++ src/HyperMC_old/MainFrame.resx
