+++ src/HyperMC_old/Mod.cs
@@ -1,0 +1,7 @@
+﻿namespace Hypermc
+{
+    public class Mod
+    {
+        public string Name { get; init; }
+    }
+}
