+++ src/HyperMC_old/ModDownloadEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Hypermc
+{
+    public class ModDownloadEventArgs : EventArgs
+    {
+        public Mod Mod { get; init; }
+        public ModDownloadEventArgs(Mod mod) => Mod = mod;
+    }
+}
